Add the most rudimentary type checking that at least validates indexing works as expected since the row size is assumed to be T.sizeof.  This may catch missing {.packed.} errors or wide row vs column errors.
If we add a more detailed checking capability here, we should also allow
a `cast` like escape hatch to be less strict.  Someone could reasonably
want to treat signed ints as unsigned, for example.  (All that really
changes is NA codes - and there may be none or the caller may not care.)