Add support for incremental maintenance of stacked matrices (time series in file terminology but any two strings could work).  This includes the types `AxisFile` & its APIs, `StacksUpdater` & its APIs and a new sub- command for users called `nio upstacks`.  (A script driven demo/vignette is inbound, as is an easy to use NumPy access module.)
Several things are manual that can & should be automated.  E.g., the <x>
in ix2tm.N<x>C must match the (fixed) length of all times discovered in
paths for the table lookups to work.  That said, it gives a useful
flavor for the kind of of data repository that can be maintained with
reasonable amortized costs.