Fix a bad bug where the wrong `$` (the generic one) was picked up.  This was easy to miss due to the `from memfiles as mf import nil` to get the qualified `mf.open` separation from `system.open`.  Things *also* worked for the simple test of loading up a file, but the keys loaded at open were wrong for `rkDelim`-style repos - like "(data: ..." ), not Nim strings.
The consequence was actually bad for .D repos - never re-pointing to any
extant string data, leading to both ever growing repos, and indirectly
also to quadratic parsing time due to the parse-to-parse file growth.
The latter is sort of always a fundamental problem for ever increasing
entropy data with any separated parsing without persistent hash tables,
but this caused that bad pattern all the time.

Add a test case to `tests/02-strings.sh` for this;  Works fine now.