BREAKING CHANGE: Output format moved to the information preserving.. backslash-escaped TSV.  Also make output easier to view/edit in a text editor by defaulting to hard-tab unit separators.
Using `-t '\0' -T "\\0"` makes things closer to the old default output
format, but still not the same.  Pretty sure breaking is necessary to
become info preserving.  So, best to do it all at once.

In the new output format only 3 things are escaped: backslash characters,
tab & nl substitutions.  You don't need to escape the subs, if you are,
e.g., destroying info on purpose (such as converting in-field delims to
space as the old defaults did).

It is possible to set flags to make output not soundly char-frameable &
split-parsable.  So, be careful.

Cost of this is about 10% perf on mid-age Intel and (a bit mysteriously)
*saves* 10-20% on AMD Zen1 under `mm:markAndSweep` (depending upon how
many cols).  Probably close to a net wash across a broad CPU testbed.

Since `markAndSweep` makes a big-ish 1.3-1.4X boost on AMD with the new
code (but not the old), add a NimScript file to turn that on.