Make able to parse only a slice of columns as numbers (or whatever) for cases like AmazonIMDB "tt0000001" parsing to just int `1` in both `load1` and in parsing schemas.
Both `cligen` `HSlice` parsers - `cligen/argcvt.argParse` for `load1`
& `cligen/mslice.parseHSlice` for parsing schemas just changed to use
a new missing upper side encodes as `U.high` convention (not `.b=0` as
it used to, to avoid ambiguity among "0:", "0:1", "0..0", "0").  So,
this code needs a bleeding edge VC head cligen to work correctly.

We use 1..0 as an "all" slice default since it is only 4 chars not the
22 of 0..9223372036854775807 which screws up help columns. Also update
`load1` help and could-not-parse logging message.